requirements.txt
Django==5.0.1
mysqlclient==2.0.1
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
kiwipiepy==0.17.0